starting refactoring
